Refactor: Reorganize section component
